requirements.txt

python-telegram-bot[job-queue]==20.7
pandas==2.2.2
openpyxl==3.1.5
tzdata==2024.1
imagehash==4.3.1
Pillow==10.4.0
